moved buildinfo package
