+++ pkg/config/provider_file.go
@@ -11,7 +11,7 @@ import (
 	"time"
 
 	"github.com/knadh/koanf/providers/file"
-	"github.com/stv0g/cunicu/pkg/util/buildinfo"
+	"github.com/stv0g/cunicu/pkg/buildinfo"
 )
 
 var (
