+++ pkg/daemon/feature/hooks/web.go
@@ -6,13 +6,13 @@ import (
 	"net"
 	"net/http"
 
+	"github.com/stv0g/cunicu/pkg/buildinfo"
 	"github.com/stv0g/cunicu/pkg/config"
 	"github.com/stv0g/cunicu/pkg/core"
 	"github.com/stv0g/cunicu/pkg/daemon/feature/epdisc"
 	icex "github.com/stv0g/cunicu/pkg/ice"
 	hooksproto "github.com/stv0g/cunicu/pkg/proto/feature/hooks"
 	rpcproto "github.com/stv0g/cunicu/pkg/proto/rpc"
-	"github.com/stv0g/cunicu/pkg/util/buildinfo"
 	"github.com/stv0g/cunicu/pkg/wg"
 	"go.uber.org/zap"
 	"golang.zx2c4.com/wireguard/wgctrl/wgtypes"
