+++ pkg/selfupdate/selfupdate_test.go
@@ -6,8 +6,8 @@ import (
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
+	"github.com/stv0g/cunicu/pkg/buildinfo"
 	"github.com/stv0g/cunicu/pkg/selfupdate"
-	"github.com/stv0g/cunicu/pkg/util/buildinfo"
 	"github.com/stv0g/cunicu/test"
 )
